+++ src/main/java/com/github/unchama/buildassist/PlayerData.java
@@ -87,7 +87,7 @@ public class PlayerData {
 
 	//建築系データを読み込む　ture:読み込み成功　false:読み込み失敗
 	public boolean buildload(Player player){
-		com.github.unchama.seichiassist.data.PlayerData playerdata_s = SeichiAssist.Companion.getPlayermap().get(uuid);
+		com.github.unchama.seichiassist.data.player.PlayerData playerdata_s = SeichiAssist.Companion.getPlayermap().get(uuid);
 		if(playerdata_s == null){
 			return false;
 		}
@@ -123,7 +123,7 @@ public class PlayerData {
 
 	//建築系データを保存
 	public void buildsave(Player player){
-		com.github.unchama.seichiassist.data.PlayerData playerData = SeichiAssist.Companion.getPlayermap().get(uuid);
+		com.github.unchama.seichiassist.data.player.PlayerData playerData = SeichiAssist.Companion.getPlayermap().get(uuid);
 		if (playerData == null){
 			player.sendMessage(ChatColor.RED+"建築系データ保存失敗しました");
 			return;
