+++ src/main/java/com/github/unchama/seichiassist/data/ActiveSkillInventoryData.java
@@ -2,6 +2,7 @@ package com.github.unchama.seichiassist.data;
 
 import com.github.unchama.seichiassist.ActiveSkill;
 import com.github.unchama.seichiassist.SeichiAssist;
+import com.github.unchama.seichiassist.data.player.PlayerData;
 import com.github.unchama.seichiassist.database.DatabaseGateway;
 import com.github.unchama.seichiassist.util.AsyncInventorySetter;
 import com.github.unchama.seichiassist.util.ItemMetaFactory;
