+++ src/main/java/com/github/unchama/seichiassist/data/RegionMenuData.java
@@ -2,6 +2,7 @@ package com.github.unchama.seichiassist.data;
 
 import com.github.unchama.seichiassist.Config;
 import com.github.unchama.seichiassist.SeichiAssist;
+import com.github.unchama.seichiassist.data.player.PlayerData;
 import com.github.unchama.seichiassist.util.Util;
 import com.github.unchama.seichiassist.util.Util.DirectionType;
 import com.github.unchama.seichiassist.util.external.ExternalPlugins;
