+++ src/main/java/com/github/unchama/seichiassist/database/manipulators/DonateDataManipulator.kt
@@ -1,7 +1,7 @@
 package com.github.unchama.seichiassist.database.manipulators
 
 import com.github.unchama.seichiassist.ActiveSkillPremiumEffect
-import com.github.unchama.seichiassist.data.PlayerData
+import com.github.unchama.seichiassist.data.player.PlayerData
 import com.github.unchama.seichiassist.database.DatabaseConstants
 import com.github.unchama.seichiassist.database.DatabaseGateway
 import com.github.unchama.seichiassist.task.recordIteration
