+++ src/main/java/com/github/unchama/seichiassist/event/SeichiLevelUpEvent.java
@@ -4,7 +4,7 @@ package com.github.unchama.seichiassist.event;
  * Created by karayuu on 2018/04/18
  */
 
-import com.github.unchama.seichiassist.data.PlayerData;
+import com.github.unchama.seichiassist.data.player.PlayerData;
 import org.bukkit.entity.Player;
 
 /**
