+++ src/main/java/com/github/unchama/seichiassist/listener/PlayerJoinListener.kt
@@ -3,7 +3,7 @@ package com.github.unchama.seichiassist.listener
 import com.github.unchama.seichiassist.ActiveSkill
 import com.github.unchama.seichiassist.ManagedWorld
 import com.github.unchama.seichiassist.SeichiAssist
-import com.github.unchama.seichiassist.data.PlayerData
+import com.github.unchama.seichiassist.data.player.PlayerData
 import com.github.unchama.seichiassist.isSeichi
 import com.github.unchama.seichiassist.util.Util
 import net.coreprotect.model.Config
