+++ src/main/java/com/github/unchama/seichiassist/listener/RegionInventoryListener.java
@@ -3,8 +3,8 @@ package com.github.unchama.seichiassist.listener;
 import com.github.unchama.seichiassist.Config;
 import com.github.unchama.seichiassist.SeichiAssist;
 import com.github.unchama.seichiassist.data.GridTemplate;
-import com.github.unchama.seichiassist.data.PlayerData;
 import com.github.unchama.seichiassist.data.RegionMenuData;
+import com.github.unchama.seichiassist.data.player.PlayerData;
 import com.github.unchama.seichiassist.util.Util;
 import com.github.unchama.seichiassist.util.Util.Direction;
 import com.github.unchama.seichiassist.util.Util.DirectionType;
