+++ src/main/java/com/github/unchama/seichiassist/listener/new_year_event/NewYearBagListener.java
@@ -2,7 +2,7 @@ package com.github.unchama.seichiassist.listener.new_year_event;
 
 import com.github.unchama.seichiassist.Config;
 import com.github.unchama.seichiassist.SeichiAssist;
-import com.github.unchama.seichiassist.data.PlayerData;
+import com.github.unchama.seichiassist.data.player.PlayerData;
 import com.github.unchama.seichiassist.util.Util;
 import com.github.unchama.seichiassist.util.external.ExternalPlugins;
 import de.tr7zw.itemnbtapi.NBTItem;
