+++ src/main/java/com/github/unchama/seichiassist/listener/new_year_event/NewYearItemListener.java
@@ -3,7 +3,7 @@ package com.github.unchama.seichiassist.listener.new_year_event;
 import com.github.unchama.seichiassist.Config;
 import com.github.unchama.seichiassist.SeichiAssist;
 import com.github.unchama.seichiassist.data.Mana;
-import com.github.unchama.seichiassist.data.PlayerData;
+import com.github.unchama.seichiassist.data.player.PlayerData;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.Material;
