+++ src/main/java/com/github/unchama/seichiassist/listener/new_year_event/NewYearsEvent.java
@@ -2,7 +2,7 @@ package com.github.unchama.seichiassist.listener.new_year_event;
 
 import com.github.unchama.seichiassist.Config;
 import com.github.unchama.seichiassist.SeichiAssist;
-import com.github.unchama.seichiassist.data.PlayerData;
+import com.github.unchama.seichiassist.data.player.PlayerData;
 import com.github.unchama.seichiassist.util.Util;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
