+++ src/main/java/com/github/unchama/seichiassist/task/CheckAlreadyExistPlayerDataTask.java
@@ -2,7 +2,7 @@ package com.github.unchama.seichiassist.task;
 
 import com.github.unchama.seichiassist.Config;
 import com.github.unchama.seichiassist.SeichiAssist;
-import com.github.unchama.seichiassist.data.PlayerData;
+import com.github.unchama.seichiassist.data.player.PlayerData;
 import com.github.unchama.seichiassist.database.DatabaseConstants;
 import com.github.unchama.seichiassist.database.DatabaseGateway;
 import org.bukkit.Bukkit;
