+++ src/main/java/com/github/unchama/seichiassist/task/GiganticBerserkTask.java
@@ -3,7 +3,7 @@ package com.github.unchama.seichiassist.task;
 import com.github.unchama.seichiassist.LevelThresholds;
 import com.github.unchama.seichiassist.SeichiAssist;
 import com.github.unchama.seichiassist.data.Mana;
-import com.github.unchama.seichiassist.data.PlayerData;
+import com.github.unchama.seichiassist.data.player.PlayerData;
 import com.github.unchama.seichiassist.util.Util;
 import net.md_5.bungee.api.ChatColor;
 import org.bukkit.Sound;
