+++ src/main/java/com/github/unchama/seichiassist/task/PlayerDataLoadTask.kt
@@ -6,8 +6,6 @@ import com.github.unchama.seichiassist.MineStackObjectList
 import com.github.unchama.seichiassist.SeichiAssist
 import com.github.unchama.seichiassist.data.GridTemplate
 import com.github.unchama.seichiassist.data.LimitedLoginEvent
-import com.github.unchama.seichiassist.data.MineStack
-import com.github.unchama.seichiassist.data.PlayerData
 import com.github.unchama.seichiassist.data.player.*
 import com.github.unchama.seichiassist.data.player.settings.BroadcastMutingSettings
 import com.github.unchama.seichiassist.database.DatabaseConstants
